requirements-dev.txt
# Install this in a Python2 virtualenv to run the tests
# Tests can be run by: python -m pytest paht/to/MrBeamPlugin/tests/

# The tags can be changed here (after @):
# Todo: after merging mrb3-doc, point to the alpha branch
git+https://github.com/mrbeam/mrb3-doc.git@feature/SW-1606-mrb-3-doc-to-python-3-10-6#egg=MrBeamDoc
git+https://github.com/mrbeam/OctoPrint.git@feature/SW-1030-octoprint-upgrade-to-latest-v-1-x#egg=OctoPrint

mock
numpy
opencv-python
Pillow
pre-commit
pytest
pytest-datafiles
pytest-mock
requests-mock
selenium
webdriver-manager
pytest-cov
pytest-html
python-statemachine